requirements.txt
fastapi
uvicorn
pymongo
pydantic
motor
python-jose
redis
passlib